+++ dakara_server/dakara_server/urls.py
@@ -18,7 +18,6 @@ from django.contrib import admin
 from django.conf import settings
 from django.contrib.staticfiles.views import serve
 from django.urls import include, path
-from rest_framework.authtoken.views import obtain_auth_token
 from rest_framework.documentation import include_docs_urls
 
 from library import views as library_views
@@ -31,18 +30,13 @@ urlpatterns = [
     # Admin route
     path("admin/", admin.site.urls),
     # Authentication routes
+    path("api/accounts/", include("rest_registration.api.urls")),
     path("api/auth/", include("rest_framework.urls", namespace="rest_framework")),
-    path("api/token-auth/", obtain_auth_token),
     # API routes for internal
-    path("api/version/", internal_views.VersionView.as_view(), name="version"),
+    path("api/settings/", internal_views.SettingsView.as_view(), name="settings"),
     # API routes for the users
     path("api/users/", users_views.UserListView.as_view(), name="users-list"),
     path("api/users/<int:pk>/", users_views.UserView.as_view(), name="users"),
-    path(
-        "api/users/<int:pk>/password/",
-        users_views.PasswordView.as_view(),
-        name="users-password",
-    ),
     path(
         "api/users/current/",
         users_views.CurrentUserView.as_view(),