+++ dakara_server/internal/views.py
@@ -5,15 +5,19 @@ from rest_framework.views import APIView
 from rest_framework.permissions import AllowAny
 
 
-class VersionView(APIView):
-    """Version of the Dakara server
+class SettingsView(APIView):
+    """Settings of the Dakara server
     """
 
     permission_classes = (AllowAny,)
 
     def get(self, request):
-        """Get application version number
+        """Get application settings, version and date
         """
-        data = {"version": settings.VERSION, "date": settings.DATE}
+        data = {
+            "version": settings.VERSION,
+            "date": settings.DATE,
+            "email_enabled": settings.EMAIL_ENABLED,
+        }
 
         return Response(data, status.HTTP_200_OK)