+++ dakara_server/users/tests/conftest.py
@@ -8,3 +8,14 @@ def users_provider():
     provider = UsersProvider()
     provider.create_test_data()
     return provider
+
+
+@pytest.fixture
+def config_email_disabled(settings):
+    rest_registration = settings.REST_REGISTRATION.copy()
+    rest_registration["REGISTER_VERIFICATION_ENABLED"] = False
+    rest_registration["REGISTER_EMAIL_VERIFICATION_ENABLED"] = False
+    rest_registration["RESET_PASSWORD_VERIFICATION_ENABLED"] = False
+
+    settings.EMAIL_ENABLED = False
+    settings.REST_REGISTRATION = rest_registration